refactor(components): refactor card
